demo/requirements.txt
anthropic>=0.3.0
requests>=2.25.0
python-dotenv>=0.19.0
../  # Install the eiondb package from parent directory 